added some authentication in some controllers

+++ server/src/controllers/postControllers.ts
@@ -1,16 +1,15 @@
 import { RequestHandler } from "express"
 import PostModel from "../models/post"
-import mongoose, { Types } from "mongoose"
+import mongoose from "mongoose"
 import createHttpError from "http-errors"
 import cloudinary from "cloudinary"
+import { CustomRequest } from "../app"
 
-type CreatePostBody = {
-    creator: string
+export type CreatePostBody = {
     content: string
-    image?: string | null
 }
 
-type UpdatePostBody = {
+export type UpdatePostBody = {
     content: string,
     image?: string
 }
@@ -19,20 +18,18 @@ type PostParam = {
     postId: string
 }
 
-type UserIdParam = {
-    userId: string
-}
-
-type LikePostBody = {
+type GetUserPostsParam = {
     userId: string
 }
 
 // create post
-export const createPost: RequestHandler<unknown, unknown, CreatePostBody, unknown> = async (req: any, res, next) => {
-    const { creator, content } = req.body
-    
-    if(!creator){
-        throw createHttpError(400, "Bad request, post should have creator")
+// NEEDS AUTHENTICATED USER IN ORDER TO ACCESS THIS
+export const createPost: RequestHandler = async (req: CustomRequest, res, next) => {
+    const { content } = req.body as CreatePostBody
+    const authenticatedCreator = req.userId
+
+    if(!authenticatedCreator){
+        throw createHttpError(403, "Forbidden, unauthorized to create a post")
     }
     
     if(!content){ 
@@ -43,14 +40,15 @@ export const createPost: RequestHandler<unknown, unknown, CreatePostBody, unknow
         let image = null;
 
         if (req.file) {
-            const imageResult: any = await cloudinary.v2.uploader.upload(req.file.path);
+            const imageResult: any = await cloudinary.v2.uploader.upload(req.file.path)
             image = {
                 url: imageResult.secure_url,
                 cloudinary_id: imageResult.public_id,
             };
         }
+        
         const newPost = await PostModel.create({
-            creator,
+            creator: authenticatedCreator,
             content,
             image
         })
@@ -64,16 +62,16 @@ export const createPost: RequestHandler<unknown, unknown, CreatePostBody, unknow
 // get all post 
 export const getPosts: RequestHandler = async (req, res, next) => {
         try {
-            const posts = await PostModel.find().populate("creator").populate("liked_by").exec();
+            // gets post in descending order
+            const posts = await PostModel.find().sort({ createdAt: -1}).populate("creator").populate("liked_by").exec();
             res.status(200).json(posts);
         } catch (error) {
             next(error)
         }
 }
 
-
 // get user's post
-export const getUserPosts: RequestHandler<UserIdParam, unknown, unknown, unknown> = async (req, res, next) => {
+export const getUserPosts: RequestHandler<GetUserPostsParam> = async (req, res, next) => {
     const userId = req.params.userId
 
     try {
@@ -94,17 +92,21 @@ export const getUserPosts: RequestHandler<UserIdParam, unknown, unknown, unknown
 } 
 
 // update post
-export const updatePost: RequestHandler<PostParam, unknown, UpdatePostBody, unknown> = async (req: any, res, next) => {
-    const  { content } = req.body
+// NEEDS AUTHENTICATED USER IN ORDER TO ACCESS THIS
+export const updatePost: RequestHandler<PostParam> = async (req: CustomRequest, res, next) => {
+    const  { content } = req.body as UpdatePostBody
     const { postId } = req.params
+    const authenticatedUserId = req.userId
     const newImageFile = req.file?.path
-    let newImageResult: any;
+    let newImageResult: any ;
 
     try {
+        let image = null
+
         if(!content){
             throw createHttpError(400, "Bad request, post should have content")
         }
-
+        
         const post = await PostModel.findById(postId).exec()
         const currentCloudinaryId = post?.image?.cloudinary_id
 
@@ -112,22 +114,37 @@ export const updatePost: RequestHandler<PostParam, unknown, UpdatePostBody, unkn
             throw createHttpError(404, "Post not found")    
         }
 
-        /* checks if there is a current image in the document and if the user has a new image, then the current image will be
-        deleted and the new image will be added */
-        if(currentCloudinaryId && newImageFile){
-            await cloudinary.v2.api.delete_resources([currentCloudinaryId]);
-            newImageResult = await cloudinary.v2.uploader.upload(newImageFile)
-        } else if(newImageFile){
-            newImageResult = await cloudinary.v2.uploader.upload(newImageFile)
-        }
+        if(authenticatedUserId){
+            if (!post.creator.equals(authenticatedUserId)) {
+                throw createHttpError(403, "Forbidden, unauthorized to delete the post");
+            }
+        } 
+        
+        // this will only work if the req has a file to upload
+        if(req.file){
+            /* checks if there is a current image in the document and if the user has a new image, then the current image will be
+            deleted and the new image will be added */
+            if(currentCloudinaryId && newImageFile){
+                await cloudinary.v2.api.delete_resources([currentCloudinaryId]);
+                newImageResult = await cloudinary.v2.uploader.upload(newImageFile)
+                image = {
+                    url: newImageResult.secure_url,
+                    cloudinary_id: newImageResult.public_id
+                }
+            } else if(newImageFile){
+                newImageResult = await cloudinary.v2.uploader.upload(newImageFile)
+                image = {
+                    url: newImageResult.secure_url,
+                    cloudinary_id: newImageResult.public_id
+                }
+            }
+        } 
 
         post.content = content
         post.image = {
-            // retains the url and cloudinary if there is no new image
-            url: newImageResult.secure_url || post.image?.url,
-            cloudinary_id: newImageResult.public_id || post.image?.cloudinary_id
+            url: image?.url || post?.image?.url,
+            cloudinary_id: image?.cloudinary_id || post?.image?.cloudinary_id
         }
-
         const updatedPost = await post.save()
 
         res.status(200).json(updatedPost);
@@ -136,16 +153,23 @@ export const updatePost: RequestHandler<PostParam, unknown, UpdatePostBody, unkn
     }
 }
 
-
 // delete post
-export const deletePost: RequestHandler<PostParam> = async (req, res, next) => {
+// NEEDS AUTHENTICATED USER IN ORDER TO ACCESS THIS
+export const deletePost: RequestHandler<PostParam> = async (req: CustomRequest, res, next) => {
     const postId = req.params.postId
+    const authenticatedUserId = req.userId
 
-    try {
+    try {     
         const post = await PostModel.findById(postId).exec()
 
-        if(!post){
-            throw createHttpError(404, "Post not found")
+        if (!post) {
+            throw createHttpError(404, "Post not found");
+        }
+
+        if(authenticatedUserId){
+            if (!post.creator.equals(authenticatedUserId)) {
+                throw createHttpError(403, "Forbidden, unauthorized to delete the post");
+            }
         }
 
         const cloudinaryId = post.image?.cloudinary_id
@@ -163,24 +187,30 @@ export const deletePost: RequestHandler<PostParam> = async (req, res, next) => {
 }
 
 // like post
-export const likePost: RequestHandler<PostParam, unknown, LikePostBody, unknown> = async (req, res, next) => {
+// NEEDS AUTHENTICATED USER IN ORDER TO ACCESS THIS
+export const likePost: RequestHandler<PostParam> = async (req: CustomRequest, res, next) => {
     const { postId } = req.params
-    const { userId } = req.body
+    const authenticatedUserId = req.userId
 
     try {  
         const post = await PostModel.findById(postId).exec()   
+
+        if(!authenticatedUserId){
+            throw createHttpError(403, "Forbidden, unauthorized to like a post")
+        }
+
         if(!post) {
             throw createHttpError(404, "Post not found")
         }
-
-        const didLike = post.liked_by.some((user) => user.equals(userId));
+        
+        const didLike = post.liked_by.some((user) => user.equals(new mongoose.Types.ObjectId(authenticatedUserId)));
     
         if(didLike){
             throw createHttpError(400, "User already liked the post")
         } 
         
         post.likes += 1
-        post.liked_by.push(new mongoose.Types.ObjectId(userId))
+        post.liked_by.push(new mongoose.Types.ObjectId(authenticatedUserId))
 
         await post.save()
 
@@ -190,19 +220,24 @@ export const likePost: RequestHandler<PostParam, unknown, LikePostBody, unknown>
     }
 }
 
-export const unlikePost: RequestHandler<PostParam, unknown, LikePostBody, unknown> = async (req, res, next) => {
+// NEEDS AUTHENTICATED USER IN ORDER TO ACCESS THIS
+export const unlikePost: RequestHandler<PostParam> = async (req: CustomRequest, res, next) => {
     const { postId } = req.params
-    const { userId } = req.body
+    const authenticatedUserId = req.userId
     
     try {  
         const post = await PostModel.findById(postId).exec()   
+        
+        if(!authenticatedUserId){
+            throw createHttpError(403, "Forbidden, unauthorized to unlike a post")
+        }
 
         if(!post) {
             throw createHttpError(404, "Post not found")
         }
         
         post.likes -= 1
-        post.liked_by = post.liked_by.filter((user) => !user.equals(userId));
+        post.liked_by = post.liked_by.filter((user) => !user.equals(new mongoose.Types.ObjectId(authenticatedUserId)));
 
         await post.save()
 
@@ -212,17 +247,23 @@ export const unlikePost: RequestHandler<PostParam, unknown, LikePostBody, unknow
     }
 }
 
-export const checkIfUserLikes: RequestHandler<PostParam, unknown, LikePostBody, unknown> = async (req, res, next) => {
+// NEEDS AUTHENTICATED USER IN ORDER TO ACCESS THIS
+export const checkIfUserLikes: RequestHandler<PostParam> = async (req: CustomRequest, res, next) => {
     const { postId } = req.params;
-    const { userId } = req.body;
+    const authenticatedUserId = req.userId
 
     try {
         const post = await PostModel.findById(postId).exec();
+
+        if(!authenticatedUserId){
+            throw createHttpError(403, "Forbidden, unauthorized to check if the user likes the post")
+        }
+        
         if (!post) {
             throw createHttpError(404, "Post not found");
         }
         
-    const didLike = post.liked_by.some((likedUserId) => likedUserId.equals(userId));
+    const didLike = post.liked_by.some((likedUserId) => likedUserId.equals(new mongoose.Types.ObjectId(authenticatedUserId)));
 
     res.json(didLike);
     } catch (error) {
